+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Find the missing plus signs in addition".
 - Tests for "Magic stones".
 - Tests for "Game of life".
+- Tests for "Maximum sub-sequence".
 
 ## [3.13.0] - 2022-09-30
 ### Added